+++ scripts/dev.config.mjs
@@ -2,7 +2,13 @@ export default {
   buildCommand: ['pnpm', 'run', 'dev:build'],
   deploy: {
     mode: 'copy',
-    staticFiles: [{ from: 'manifest.json', to: 'manifest.json' }],
-    watchFiles: [{ from: 'main.js', to: 'main.js' }],
+    staticFiles: [
+      { from: 'manifest.json', to: 'manifest.json' },
+      { from: 'styles.css', to: 'styles.css' },
+    ],
+    watchFiles: [
+      { from: 'main.js', to: 'main.js' },
+      { from: 'styles.css', to: 'styles.css' },
+    ],
   },
 }